Issue 38: Put admin actions in "admin" controllers
