+++ src/Khatovar/Bundle/WebBundle/Resources/config/routing/appearance.yml
@@ -10,39 +10,39 @@ khatovar_web_appearance_camp:
     path:     /campement
     defaults: { _controller: KhatovarWebBundle:Appearance:camp }
 
-khatovar_web_appearance_list:
-    path:     /list
-    defaults: { _controller: KhatovarWebBundle:Appearance:list }
-
 khatovar_web_appearance_show:
     path:     /prestation/{slug}
     defaults: { _controller: KhatovarWebBundle:Appearance:show }
     requirements:
         slug: '[a-zA-Z0-9-_/]+'
 
+khatovar_web_appearance_list:
+    path:     /list
+    defaults: { _controller: KhatovarWebBundle:Admin\Appearance:list }
+
 khatovar_web_appearance_new:
     path:     /new
-    defaults: { _controller: KhatovarWebBundle:Appearance:new }
+    defaults: { _controller: KhatovarWebBundle:Admin\Appearance:new }
 
 khatovar_web_appearance_create:
     path:     /create
-    defaults: { _controller: KhatovarWebBundle:Appearance:create }
+    defaults: { _controller: KhatovarWebBundle:Admin\Appearance:create }
     methods:  POST
 
 khatovar_web_appearance_edit_index:
     path:     /index/edit
-    defaults: { _controller: KhatovarWebBundle:Appearance:editIndex }
+    defaults: { _controller: KhatovarWebBundle:Admin\Appearance:editIndex }
 
 khatovar_web_appearance_edit:
     path:     /{id}/edit
-    defaults: { _controller: KhatovarWebBundle:Appearance:edit }
+    defaults: { _controller: KhatovarWebBundle:Admin\Appearance:edit }
 
 khatovar_web_appearance_update:
     path:     /{id}/update
-    defaults: { _controller: KhatovarWebBundle:Appearance:update }
+    defaults: { _controller: KhatovarWebBundle:Admin\Appearance:update }
     methods:  [POST, PUT]
 
 khatovar_web_appearance_delete:
     path:     /{id}/delete
-    defaults: { _controller: KhatovarWebBundle:Appearance:delete }
+    defaults: { _controller: KhatovarWebBundle:Admin\Appearance:delete }
     methods:  [POST, DELETE]
