+++ src/Khatovar/Bundle/WebBundle/Resources/config/routing/camp.yml
@@ -1,3 +1,3 @@
 khatovar_web_camp:
     path:     /
-    defaults: { _controller: KhatovarWebBundle:Web:camp }
+    defaults: { _controller: KhatovarWebBundle:Camp:camp }
